Open Class RemoteService
